Turn some i16 types to u16

+++ src/item_list_controller.rs
@@ -73,7 +73,7 @@ impl<'a, T> ItemListController<'a, T> where T: Write {
         let (display_width, _) = size()?;
         self.item_list_model.reset_selection_window_height();
         let count = self.item_list.get_renderable_items_count(display_width, self.item_list_model);
-        self.item_list_model.set_selection_window_height(count as i16);
+        self.item_list_model.set_selection_window_height(count);
         self.item_list.refresh(display_width, self.item_list_model)
     }
